Add files via upload
This is a small project of mine, This project mainly defines the Percentage and grade a student gets from his/her marks in their Exam. It takes the input of 6 subjects, the Grand totals of the marks, and then find the percentage. Then with this percentage, we use a switch statement for finding the grade.